System: add additional user session checks on root-level actions

+++ notificationsDeleteAllProcess.php
@@ -25,6 +25,11 @@ include './gibbon.php';
 
 $URL = Url::fromRoute('notifications')->withQueryParam('sidebar', 'false');
 
+if (!$session->has('gibbonPersonID') || !$session->has('gibbonRoleIDCurrent')) {
+    header("Location: {$URL->withReturn('error0')}");
+    exit;
+}
+
 try {
     $data = array('gibbonPersonID' => $session->get('gibbonPersonID'));
     $sql = 'DELETE FROM gibbonNotification WHERE gibbonPersonID=:gibbonPersonID';
